[sdk] move TextSlide to /views
