+++ sdk/views/TextSlide.tsx
